+++ app/config.js
@@ -5,7 +5,7 @@ var url = require('url');
 var coins = require("./coins.js");
 var credentials = require("./credentials.js");
 
-var currentCoin = process.env.BTCEXP_COIN || "BCH";
+var currentCoin = process.env.BTCEXP_COIN || "TRRXITTE";
 
 var rpcCred = credentials.rpc;
 
@@ -184,8 +184,8 @@ module.exports = {
             // TODO: find better images for this URLs, in the mean time use bch.svg
             {name: "Bitcoin Unlimited", url:"https://www.bitcoinunlimited.info", imgUrl:"/img/logo/bu.png"},
             {name: "Bitcoin Cash Nodes", url:"https://cashnodes.io", imgUrl:"/img/logo/bch.svg"},
-            {name: "BCH PoW Monitoring", url:"https://fork.lol", imgUrl:"/img/logo/fork.png"},
-            {name: "BCH Specification", url:"https://reference.cash", imgUrl:"/img/logo/refcash.ico"},
+            {name: "TRRXITTE PoW Monitoring", url:"https://fork.lol", imgUrl:"/img/logo/fork.png"},
+            {name: "TRRXITTE Specification", url:"https://reference.cash", imgUrl:"/img/logo/refcash.ico"},
             {name: "Bitcoin Cash Node", url:"https://bitcoincashnode.org/", imgUrl:"/img/logo/bchn.png"},
           ]
         }
@@ -201,10 +201,10 @@ module.exports = {
 
   donations:{
     addresses:{
-      coins:["BCH"],
-      sites:{"BCH":"https://www.bitcoinunlimited.info"},
+      coins:["TRRXITTE"],
+      sites:{"TRRXITTE":"https://www.bitcoinunlimited.info"},
 
-      "BCH":{address:"bitcoincash:pq6snv5fcx2fp6dlzg7s0m9zs8yqh74335tzvvfcmq"}
+      "TRRXITTE":{address:"bitcoincash:pq6snv5fcx2fp6dlzg7s0m9zs8yqh74335tzvvfcmq"}
     }
   }